# PERF: normalize_i8_timestamps