xref #34507

Based on the asv mentioned in that thread:

```
In [3]: N = 10**5                                                                                                                                                                                              
In [4]: tz = "UTC"                                                              
In [5]: series = Series(date_range(start="1/1/2000", periods=N, freq="T", tz=tz))

In [6]: %timeit series.dt.normalize()                                                                                                                                                                          
947 µs ± 4.33 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)   # <-- PR
3.26 ms ± 43.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)   # <-- master
```